Fixed issue that was causing parameters to not be saved to file in DataSurface, so old surfaces were not eing recognized even though they were the same.

+++ cratermaker/utils/general_utils.py
@@ -519,4 +519,4 @@ def convert_deg2m_res(inres: FloatLike, radius: FloatLike) -> float:
     -------
     float : The converted value of either m/px or px/deg, depending on which one was input
     """
-    return np.pi / 180.0 * radius / inres
+    return float(np.pi / 180.0 * radius / inres)
